Propagate pool-id and pool-name logging
Previously, the pool_id and pool_name log values were not propagated to
connection-level log messages. This change sets the pool_id and pool_name
in the thread environment, so that those variables are available farther
down the call stack.

Additonally, the pool test suite was not being run due to a typo.
As a bonus, this change fixes the pool tests to run.